Refactor: renamed db to upload
